Get rid of sysexit (#51)
Replaced sys.exit with raising an appropiate exception.

`sys.exit()` was used to quit upon an invalid state. This is very misleading, since `sys.exit()` returns UNIX error code 0 and is not considered to be an erroneous execution by e.g. the test-suite. 
One could call `sys.exit(1)` but even better is to properly replace with exceptions.

Fixes #12