Unified fragpart (#131)
Implemented a clear separation between `FragPart`, the class that holds the fragment data, and `fragmentate` (previously `fragpart`), the function that computes it.

Now `FragPart` documents its components and `fragpart` documents its arguments. (These two are not the same!)

This was done for both `kbe` and `molbe`.

This more explicitly shows which components the three fragmentation codes `autogen`, `graphgen`, and `chemgen` (have) to support. Relevant for #112 

Renamed all components that come out of  `autogen`, `graphgen`, and `chemgen`  in a consistent way.

------------------------------------------------

Additionally I removed `polychain` from the `kbe` `fragmentate`. It  was buggy and unused.