Faster absolute overlap (#177)
Improve the speed of the calculation of the absolute overlap, but uses **approximate** absolute overlaps.

Compare with
https://github.com/pyscf/pyscf/pull/2913

The current way to compute the absolute overlap is only exact for uncontracted, cartesian basis functions.
As soon as one uses contracted cartesians or uses spherical harmonics it is necessary to transform the overlap matrix from the primitives.

This can be done only in an approximated way using the RHS of the triangle inequality:

```math
\int |\phi_i(\mathbf{r})| \, |\phi_j(\mathbf{r})| \, d\mathbf{r}
\leq
\sum_{\alpha,\beta} |c_{\alpha i}| \, |c_{\beta j}| \int |\chi_\alpha(\mathbf{r})| \, |\chi_\beta(\mathbf{r})| \, d\mathbf{r}
```