Fix skipped tests (#196)
Fixes #195 

Fixes currently skipped tests and ensures we don't run into this issue anymore.

Tests the whole test suite, including expensive tests, after `main` changed. 
This way the expensive tests are not executed after every push to a PR, but we also don't run blind for months.

Makes it easier for users to just run `pytest`, because one has to explicitly specify to **run** the expensive tests; instead of explicitly specifying to **skip** the tests as it was before.

In addition it uses `skipUnless(condition)` instead of `skipIf(not condition)` to avoid double negations.