Remove numba sparse DF (#236)
Removed numba implementation of sparse DF.

This simplified the code as it removes a lot of unused functionality.

Furthermore we can use the C++ objects exposed to python now more freely.
Reduced the memory usage by 1/3 by doing manual allocations on the C++ side instead of passing a copy of the numpy array.

It also removes the screening based on atom distance, as the screening based on absolute overlap of the orbtials is much more efficient.
I will tag this PR with "removed functionality" as this might perhaps become relevant again.